SWF-17305 - fixing new Xcode 9 build warning
Xcode 9 was generating the following warning:

PBJVision.m:762:61: warning: null passed to a callee that requires a non-null argument [-Wnonnull]
        _previewLayer = [[AVCaptureVideoPreviewLayer alloc] initWithSession:nil];
                                                            ^               ~~~

I hacked in a simple, minimal fix, to keep from diverging from the original repo as much as possible. Instead of passing in nil, I pass in a local variable which is set to nil. Looks super ugly but it's good enough to shut the compiler up.

CR: Jon Moldover